% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{densitycut.clust}
\alias{densitycut.clust}
\title{DensityCut caller}
\usage{
densitycut.clust(input_CpG_data, mean_meth_matrix, R, max_PC = 20,
  maxit = 100, impute, use_cache, outdir)
}
\arguments{
\item{input_CpG_data}{A matrix of methylation data}

\item{mean_meth_matrix}{A matrix of mean methylation per region}

\item{R}{An integer, number of regions}

\item{max_PC}{An integer, max number of PC components}

\item{maxit}{An integer, max number of iterations for DensityCut}

\item{impute}{if 1, fill in NAs with imputed data, if 0, do nothing}

\item{use_cache}{if 1, use cached data when available, if 0, recompute data}

\item{outdir}{file path to output directory}
}
\value{
Predicted clusters from DensityCut algorithm
}
\description{
DensityCut caller
}
\examples{
data <- load_data(input_CpG_data_file, input_regions_file)

input_CpG_data <- data$input_CpG_data
input_regions <- data$input_regions

R <- dim(input_regions)[1] ## number of regions
maxpc <- min(20, R)
outdir <- args$outdir

densitycut.clust(input_CpG_data, mean_meth_matrix, R, max_PC, maxit=100, impute=0, use_cache=1, outdir)

}
