% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{euclidean.clust}
\alias{euclidean.clust}
\title{Euclidean Clustering}
\usage{
euclidean.clust(input_CpG_data, mean_meth_matrix, R, Max_K,
  index_type = "ch", impute, use_cache, outdir)
}
\arguments{
\item{input_CpG_data}{A matrix of methylation data}

\item{mean_meth_matrix}{A matrix of mean methylation per region}

\item{R}{An integer, number of regions}

\item{Max_K}{An integer, maximum number of clusters to be considered when cutting the tree}

\item{index_type}{"ch" or "gap", Index to be used to choose number of clusters}

\item{impute}{if 1, impute with the average per region/locus, if it is 0 do nothing}

\item{use_cache}{if 1, use cached data when available, if 0, recompute data}

\item{outdir}{file path to output directory}
}
\value{
Predicted clusters from Euclidean clustering algorithm as a dataframe object (cell x cluster pred.)
}
\description{
Clusters methylation data based on Euclidean clustering algorithm. If R == 1, CpG based clustering is performed, if R > 1, region based clustering is performed. Also creates output files of clusters and plots in specified output directory.
}
\examples{
data <- load_data(input_CpG_data_file, input_regions_file)

input_CpG_data <- data$input_CpG_data
input_regions <- data$input_regions

R <- dim(input_regions)[1] ## number of regions
Max_K <- args$Max_K
outdir <- args$outdir

euclidean.clusters <- euclidean.clust(input_CpG_data, mean_meth_matrix, R, Max_K, index_type="ch", impute=0, use_cache=1, outdir)

}
\seealso{
\code{\link{REpiclomal}}
}
