% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_epiclomal.R
\name{evaluate.epiclomal}
\alias{evaluate.epiclomal}
\title{Evaluate epiclomal}
\usage{
evaluate.epiclomal(input, output, model, flag, criterion, GAIN_THRESHOLD)
}
\arguments{
\item{input}{Input directory of epiclomal results}

\item{output}{Output directory of evaluation results}

\item{model}{A name for the model}

\item{flag}{flag}

\item{criterion}{criterion}

\item{GAIN_THRESHOLD}{gain threshold}
}
\value{
List of best row and table best. as best_row and table_best
}
\description{
Evaluation function for epiclomal results
}
\examples{
evaluation <- evaluate.epiclomal (input, output, model, "all", "DIC_LINE_ELBOW", "0.05_-100")  # essentially means the elbow curve can go all the way to the end
best_row <- evaluation$best_row
table_best <- evaluation$table_best

}
\seealso{
\code{\link{REpiclomal}}
}
