% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hamming.clust}
\alias{hamming.clust}
\title{Hamming Clustering}
\usage{
hamming.clust(input_CpG_data, Max_K, index_type = "ch", impute,
  use_cache, outdir)
}
\arguments{
\item{input_CpG_data}{A matrix of methylation data}

\item{Max_K}{An integer, maximum number of clusters to be considered when cutting the tree}

\item{index_type}{"ch" or "gap", Index to be used to choose number of clusters. Default "ch"}

\item{impute}{if 1, impute with the average per region/locus, if it is 0 do nothing}

\item{use_cache}{if 1, use cached data when available, if 0, recompute data}

\item{outdir}{file path to output directory}
}
\value{
Predicted clusters from Euclidean clustering algorithm
}
\description{
Hamming Clustering
}
\examples{
data <- load_data(input_CpG_data_file, input_regions_file)

input_CpG_data <- data$input_CpG_data
Max_K <- args$Max_K
outdir <- args$outdir

hamming.clust(input_CpG_data, Max_K, index_type="ch", impute=0, use_cache=1, outdir)

}
