% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{load_data}
\alias{load_data}
\title{Load methylation data}
\usage{
load_data(input_CpG_data_file, input_regions_file)
}
\arguments{
\item{input_CpG_data_file}{A file containing methylation data}

\item{input_regions_file}{A file containing regions for region based clustering}
}
\value{
A list containing the parsed methylation data as a matrix (input_CpG_data),
the input regions (input_regions), and the mean methylation per region as a matrix (mean_meth_matrix)
}
\description{
Load methylation data
}
\examples{
data <- load_data(input_CpG_data_file, input_regions_file)

input_CpG_data <- data$input_CpG_data
input_regions <- data$input_regions
mean_meth_matrix <- data$mean_meth_matrix
rm(data)

}
\seealso{
\code{\link{REpiclomal}}
}

