% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_epiclomal.R
\name{hamming.dist}
\alias{hamming.dist}
\title{Hamming Distance}
\usage{
hamming.dist(outfile_est, true_epigenotype_file, true_Z_file, estimated_epigenotype_file, estimated_Z_file, methylation_file, regions_file)
}
\arguments{
\item{outfile_est}{Path to the output file for the Bayesian estimates}

\item{true_epigenotype_file}{Path to true epigenotypes}

\item{true_Z_file}{Path to true cell membership}

\item{estimated_epigenotype_file}{Path to estimated epigenotypes}

\item{estimated_Z_file}{Path to estimated cell membership}

\item{methylation_file}{Path to the methylation file}

\item{regions_file}{Path to the region file}
}
\value{
    List containing raw hamming distances, corrected inferred states, and naive imputation.
}
\description{
Calculate hamming distance between true and estimated epigenotype/clusters. Also does the adjustment of the Epiclomal inferred states, and calculates the naive imputation.
Distances are written to specified output file.
}
\examples{
hamming_dist <- hamming.dist(houtfile, true_epigenotypes_file, true_clusters_file, epiMAPfile, clMAPfile, meth_file, regions_file)
estimates <- hamming_dist$estimates
corrected_estimates <- hamming_dist$estimates
naive_data <- hamming_dist$naive_data

}
\seealso{
\code{\link{REpiclomal}}
}
