% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{visualization}
\alias{visualization}
\title{Visualization}
\usage{
hamming.clust(input_CpG_data, Max_K, index_type = "ch", impute,
  use_cache, outdir)
}
\arguments{
\item{out_dir}{Path to output directory}

\item{input_CpG_data_file}{Path to methylation data}

\item{input_CN_data_file}{Path to copy number data if available}

\item{input_regions_file}{Path to region coordinates}

\item{inferred_clusters_file}{Path to inferred clusters if available. The third column has posterior probabilities.}

\item{true_clusters_file}{Path to true clusters if available}

\item{order_by_true}{If 1, rows are ordered by the true clustering if given, else by the predicted}

\item{name}{A name for the final plot}

\item{regions_to_plot}{A file with which regions to plot, for example the flipped_regions.tsv file from the generator.}

\item{use_cache}{if 1, use cached data when available, if 0, recompute data}
}
\description{
Visualization for Epiclomal
}
\examples{
visualization(out_dir, input_CpG_data_file, input_CN_data_file, input_regions_file, inferred_clusters_file, true_clusters_file, order_by_true, name, regions_to_plot, use_cache)
}
\seealso{
\code{\link{REpiclomal}}
}

